project19/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
shapely>=1.7.0
Pillow>=8.0.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
seaborn>=0.11.0
opencv-python>=4.5.0
pytesseract>=0.3.8
easyocr>=1.6.0
scipy>=1.7.0
